AddNewProductVC Final
