+++ TimeToCook/Modules/AddingNewProduct/SubViews/CloseButtonAdd.swift
@@ -1,5 +1,5 @@
 //
-//  ContentScroll.swift
+//  AddCloseButton.swift
 //  TimeToCook
 //
 //  Created by Никита Гвоздиков on 07.08.2021.
@@ -7,17 +7,19 @@
 
 import UIKit
 
-final class ContentScroll: UIScrollView {
+final class CloseButtonAdd: UIButton {
+    
+    //MARK: Init
     
     init() {
         super.init(frame: .zero)
         translatesAutoresizingMaskIntoConstraints = false
-        keyboardDismissMode = .interactive
-        bounces = false
-        showsVerticalScrollIndicator = false
+        setImage(UIImage(systemName: "xmark"), for: .normal)
+        tintColor = #colorLiteral(red: 0.5570600033, green: 0.5567737818, blue: 0.5772830844, alpha: 1)
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
 }
+
