+++ TimeToCook/Modules/AddingNewProduct/SubViews/ContentScroll.swift
@@ -1,5 +1,5 @@
 //
-//  AddCloseButton.swift
+//  ContentScroll.swift
 //  TimeToCook
 //
 //  Created by Никита Гвоздиков on 07.08.2021.
@@ -7,19 +7,19 @@
 
 import UIKit
 
-final class AddCloseButton: UIButton {
+final class ContentScroll: UIScrollView {
     
     //MARK: Init
     
     init() {
         super.init(frame: .zero)
         translatesAutoresizingMaskIntoConstraints = false
-        setImage(UIImage(systemName: "xmark"), for: .normal)
-        tintColor = #colorLiteral(red: 0.5570600033, green: 0.5567737818, blue: 0.5772830844, alpha: 1)
+        keyboardDismissMode = .interactive
+        bounces = false
+        showsVerticalScrollIndicator = false
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
 }
-
