+++ TimeToCook/Modules/AddingNewProduct/SubViews/DownButtonForKB.swift
@@ -9,6 +9,8 @@ import UIKit
 
 final class DownButtonForKB: UIBarButtonItem {
     
+    //MARK: Init
+    
     override init() {
         super.init()
         tintColor = .white
