+++ TimeToCook/Modules/AddingNewProduct/SubViews/UpButtonForKB.swift
@@ -9,6 +9,8 @@ import UIKit
 
 final class UpButtonForKB: UIBarButtonItem {
     
+    //MARK: Init
+    
     override init() {
         super.init()
         tintColor = .white
