+++ TimeToCook/Modules/ProductInfo/SubViews/MascotImageView.swift
