+++ TimeToCook/Supporting Files/Info.plist
@@ -22,6 +22,12 @@
 	<string>1</string>
 	<key>LSRequiresIPhoneOS</key>
 	<true/>
+	<key>NSCameraUsageDescription</key>
+	<string>Для сканирования продукта, необходимо разрешить доступ к камере</string>
+	<key>Required background modes </key>
+	<array>
+		<string>App processes data in the background</string>
+	</array>
 	<key>UIApplicationSceneManifest</key>
 	<dict>
 		<key>UIApplicationSupportsMultipleScenes</key>
@@ -55,12 +61,6 @@
 	<array>
 		<string>UIInterfaceOrientationPortrait</string>
 	</array>
-	<key>Required background modes </key>
-	<array>
-		<string>App processes data in the background</string>
-	</array>
-	<key>NSCameraUsageDescription</key>
-	<string>Для сканирования продукта, необходимо разрешить доступ к камере</string>
 	<key>UISupportedInterfaceOrientations~ipad</key>
 	<array>
 		<string>UIInterfaceOrientationPortrait</string>
@@ -68,5 +68,7 @@
 		<string>UIInterfaceOrientationLandscapeLeft</string>
 		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
+	<key>UIUserInterfaceStyle</key>
+	<string>Light</string>
 </dict>
 </plist>
